requirements.txt
fastapi
pymediainfo 
python-dotenv
uvicorn
python-dotenv
